prefer studio mcp entry
